requirements.txt
streamlit
requests
pandas
Pillow

